feat: show validation errors
